Adds gcc-toolset-12 explicitly in the packages to install
If we don't add it explicitly, at the time of writing, the following packages are installed as a dependency of go-toolset:

- gcc-11.3.1-2.1.el9
- gcc-toolset-12-binutils-gold-2.38-14.el9
- gcc-toolset-12-binutils-2.38-16.el9

The build number in gcc-toolset-12-binutils-gold is 14, while in gcc-toolset-12-binutils is 16.
This brings some golang projects not to build the binary correctly (happening on arm64 with oc at the linking phase, not on amd64).
`yum upgrade` updates gcc-toolset-12-binutils-gold to build number 16.

Installing binutils in a cs9 container currently reproduces the installation of the two packages with different build numbers.