Install nodejs via nvm
The current console code cannot build on the recent versions of nodejs available in the cs9 official repositories.
Until the code base is updated to build with it, keep the nodejs version to the same one used in the rhel8 builder image currently used for the OCP builds.

Refers webpack/webpack#14532